# contracts/openapi/aihub-v1.yaml
...
openapi: 3.1.0
...
info:
  title: AIHub Platform API
  version: v1
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    内部 AI 资产管理平台 REST API 契约。统一成功响应为 ApiResponse，统一失败响应为 ApiError。
    Agent 不应依赖本地化 message 做分支判断，只依赖稳定的 code 与 retryable。
servers:
  - url: /api/v1
    description: 统一 API 前缀
...
security:
  - bearerAuth: []
paths:
...
  /system/audit-logs:
    get:
      operationId: listAuditLogs
      summary: 游标查询不可变审计记录
      tags: [audit]
      x-implementation-status: planned-p0-b
      x-required-permissions: [audit:read]
      parameters:
        - { name: principalId, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: resourceId, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: 审计记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CursorPageAuditLog' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
...
  /assets:
    post:
      operationId: createAsset
      summary: 登记新资产（模型/数据集）
      description: 校验坐标唯一、开通 Git 仓库并写入初始卡片，返回资产详情。
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:create]
      x-audit-event: ASSET_CREATED
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AssetView'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
    get:
      operationId: searchAssets
      summary: 检索资产摘要（游标分页）
      description: 精确字段过滤 + 关键词全文 + 权限可见性过滤；默认排除归档资产。
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:read]
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { $ref: '#/components/schemas/AssetType' } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
        - { name: task, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { type: string } }
        - { name: modality, in: query, schema: { type: string } }
        - { name: tagId, in: query, schema: { type: string, pattern: '^tag_' } }
        - { name: owner, in: query, schema: { type: string } }
        - { name: includeArchived, in: query, schema: { type: boolean, default: false } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: 检索成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CursorPageAssetSummary'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /assets/{assetId}:
    parameters:
      - name: assetId
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getAsset
      summary: 查询资产详情
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:read]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AssetView'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: updateAsset
      summary: 更新资产可变元数据
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:update]
      x-audit-event: ASSET_UPDATED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AssetView'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteAsset
      summary: 逻辑删除资产
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:delete]
      x-audit-event: ASSET_DELETED
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /models:
    get:
      operationId: searchModels
      summary: 检索模型列表（固定类型语义入口）
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:read]
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
        - { name: task, in: query, schema: { type: string } }
        - { name: tagId, in: query, schema: { type: string, pattern: '^tag_' } }
        - { name: owner, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: 检索成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CursorPageAssetSummary'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /datasets:
    get:
      operationId: searchDatasets
      summary: 检索数据集列表（固定类型语义入口）
      tags: [assets]
      x-implementation-status: partial-p1-frozen
      x-required-permissions: [asset:read]
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { type: string } }
        - { name: modality, in: query, schema: { type: string } }
        - { name: tagId, in: query, schema: { type: string, pattern: '^tag_' } }
        - { name: owner, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: 检索成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CursorPageAssetSummary'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        平台签发的短期访问 JWT。JWT 使用非对称签名和 kid；资源级权限由后端实时判断，
        不以 JWT 内角色替代 RBAC/ACL/资源状态校验。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 当前 Principal 范围内唯一的写请求幂等键。
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    ErrorResponse:
      description: 统一失败响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiResponse:
      type: object
      description: 统一成功响应包装
      required:
        - requestId
        - traceId
        - timestamp
      properties:
        data:
          description: 业务数据载荷，结构由具体接口定义
          type: [object, array, string, number, integer, boolean, 'null']
        requestId:
          type: string
          description: 请求标识，回显 X-Request-Id
          example: req_01J0000000000000000000000
        traceId:
          type: string
          description: 分布式追踪标识
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间戳
          example: '2026-06-29T08:00:00Z'
    ApiError:
      type: object
      description: 统一失败响应
      required:
        - code
        - message
        - retryable
        - requestId
        - traceId
      properties:
        code:
          type: string
          description: 稳定的语义错误码，分支判断只依赖此字段
          example: ASSET_VERSION_CONFLICT
        message:
          type: string
          description: 默认（可本地化）错误消息，不用于分支判断
          example: Version already exists
        i18nKey:
          type: string
          description: 国际化文案 Key
          example: error.asset.versionConflict
        details:
          type: object
          description: 附加上下文，结构由错误码定义
          additionalProperties: true
        retryable:
          type: boolean
          description: 是否可重试；仅当为 true 时 Agent 才应按退避策略重试
          example: false
        requestId:
          type: string
          example: req_01J0000000000000000000000
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
...
    AuditLogView:
      type: object
      required: [auditId, principalId, action, resourceType, resourceId, result, occurredAt]
      properties:
        auditId: { type: string, pattern: '^aud_' }
        principalId: { type: string }
        action: { type: string }
        resourceType: { type: string }
        resourceId: { type: string }
        result:
          type: string
          enum: [SUCCEEDED, FAILED, DENIED]
        errorCode: { type: [string, 'null'] }
        requestId: { type: [string, 'null'] }
        traceId: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
    CursorPageAuditLog:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuditLogView' }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
...
    AssetType:
      type: string
      description: 资产类型稳定枚举
      enum: [MODEL, DATASET]
    Visibility:
      type: string
      description: 可见性稳定枚举；PUBLIC 表示所有已认证主体可见，不代表互联网匿名访问。
      enum: [PRIVATE, INTERNAL, PUBLIC]
    AssetStatus:
      type: string
      description: 资产目录条目状态稳定枚举
      enum: [ACTIVE, DEPRECATED, ARCHIVED]
    ModelProfile:
      type: object
      description: 模型画像
      properties:
        framework:
          type: string
          description: model_framework 字典的有效 itemCode。
          example: pytorch
        task:
          type: string
          description: model_task 字典的有效 itemCode。
          example: text-generation
        architecture: { type: string }
    DatasetProfile:
      type: object
      description: 数据集画像
      properties:
        format:
          type: string
          description: dataset_format 字典的有效 itemCode。
          example: parquet
        modality:
          type: string
          description: dataset_modality 字典的有效 itemCode。
          example: text
...
    CreateAssetRequest:
      type: object
      additionalProperties: false
      required: [type, namespace, name, visibility]
      properties:
        type: { $ref: '#/components/schemas/AssetType' }
        organizationId:
          type: string
          description: 所属组织 ID
        projectId:
          type: string
          description: 所属项目 ID
        namespace:
          type: string
          description: 命名空间（小写 slug）
          example: nlp
        name:
          type: string
          description: 名称（小写 slug），namespace+type+name 唯一
          example: qwen-domain-7b
        displayName: { type: string, example: 领域问答模型 }
        description: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        owners:
          type: array
          uniqueItems: true
          description: 受权 Principal/Team 的稳定 ID。
          items: { type: string, pattern: '^(prn|team)_' }
          example: [team_01J0000000000000000000000]
        tagIds:
          type: array
          uniqueItems: true
          description: 已启用且当前主体有权使用的平台/组织受控标签 ID。
          items: { type: string, pattern: '^tag_' }
          example: [tag_01J0000000000000000000000]
        license:
          type: string
          description: license_catalog 字典的有效 itemCode。
          example: Apache-2.0
        model: { $ref: '#/components/schemas/ModelProfile' }
        dataset: { $ref: '#/components/schemas/DatasetProfile' }
    UpdateAssetRequest:
      type: object
      description: 更新可变元数据；坐标（namespace/type/name）不可变
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
          description: 资产乐观锁版本，冲突返回 ASSET_CONCURRENT_MODIFICATION。
        organizationId:
          type: string
          description: 所属组织 ID（为空表示不变更）
        projectId:
          type: string
          description: 所属项目 ID（为空表示不变更）
        displayName: { type: string }
        description: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        owners:
          type: array
          uniqueItems: true
          items: { type: string, pattern: '^(prn|team)_' }
        tagIds:
          type: array
          uniqueItems: true
          items: { type: string, pattern: '^tag_' }
        license:
          type: string
          description: license_catalog 字典的有效 itemCode。
        model: { $ref: '#/components/schemas/ModelProfile' }
        dataset: { $ref: '#/components/schemas/DatasetProfile' }
    AssetView:
      type: object
      description: 资产详情视图（非持久化实体）
      required: [assetId, type, namespace, name, visibility, status]
      properties:
        assetId: { type: string, example: ast_01J... }
        type: { $ref: '#/components/schemas/AssetType' }
        namespace: { type: string }
        organizationId: { type: [string, 'null'], description: 所属组织 ID }
        projectId: { type: [string, 'null'], description: 所属项目 ID }
        name: { type: string }
        displayName: { type: string }
        description: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        status: { $ref: '#/components/schemas/AssetStatus' }
        owners:
          type: array
          items: { type: string }
        tags:
          type: array
          description: legacy 自由标签（只读回显）
          items: { type: string }
        tagIds:
          type: array
          description: 受控标签 ID 列表（asset_tag 关联）
          items: { type: string, pattern: '^tag_' }
        license: { type: string }
        model:
          oneOf:
            - $ref: '#/components/schemas/ModelProfile'
            - type: 'null'
        dataset:
          oneOf:
            - $ref: '#/components/schemas/DatasetProfile'
            - type: 'null'
        repository:
          type: [object, 'null']
          properties:
            fullName: { type: string }
            htmlUrl: { type: string }
            cloneUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssetSummary:
      type: object
      description: 资产检索摘要（小体积）
      required: [assetId, type, namespace, name, visibility, status]
      properties:
        assetId: { type: string }
        type: { $ref: '#/components/schemas/AssetType' }
        namespace: { type: string }
        organizationId: { type: [string, 'null'], description: 所属组织 ID }
        projectId: { type: [string, 'null'], description: 所属项目 ID }
        name: { type: string }
        displayName: { type: string }
        description: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        status: { $ref: '#/components/schemas/AssetStatus' }
        owners:
          type: array
          items: { type: string }
        tags:
          type: array
          description: legacy 自由标签（只读回显）
          items: { type: string }
        tagIds:
          type: array
          description: 受控标签 ID 列表（asset_tag 关联）
          items: { type: string, pattern: '^tag_' }
        license: { type: string }
        framework: { type: string }
        task: { type: string }
        format: { type: string }
        modality: { type: string }
        updatedAt: { type: string, format: date-time }
    CursorPageAssetSummary:
      type: object
      description: 资产摘要游标分页
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AssetSummary' }
        nextCursor:
          type: [string, 'null']
          description: 下一页游标，null 表示末页
        hasMore:
          type: boolean